Save/load window state for the app windows only
